Admin pages code moved to screens folder
